chore: move electron components into plugin-electron-components
Fixes #4972
