+++ plugins/plugin-electron-components/.gitignore
@@ -1,0 +1,1 @@
+dist
