+++ plugins/plugin-electron-components/tsconfig.json
@@ -2,8 +2,8 @@
   "extends": "../../packages/builder/tsconfig-base.json",
   "include": ["src/**/*"],
   "compilerOptions": {
-    "noImplicitAny": true,
     "composite": true,
+    "jsx": "react",
     "outDir": "mdist",
     "rootDir": "src"
   }
