+++ plugins/plugin-electron-components/web/scss/components/Screenshot/Screenshot.scss
