+++ plugins/plugin-electron-components/web/scss/components/Search/Search.scss
