--- plugins/plugin-sandbox/.gitignore
@@ -1,3 +1,0 @@
-dist/
-mdist/
-build
