--- plugins/plugin-sandbox/CHANGELOG.md
@@ -1,181 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [8.10.3](https://github.com/IBM/kui/compare/v8.10.2...v8.10.3) (2020-06-25)
-
-**Note:** Version bump only for package @kui-shell/plugin-sandbox
-
-## [8.10.2](https://github.com/IBM/kui/compare/v8.10.1...v8.10.2) (2020-06-24)
-
-**Note:** Version bump only for package @kui-shell/plugin-sandbox
-
-## [8.10.1](https://github.com/IBM/kui/compare/v8.10.0...v8.10.1) (2020-06-18)
-
-**Note:** Version bump only for package @kui-shell/plugin-sandbox
-
-# [8.10.0](https://github.com/IBM/kui/compare/v4.5.0...v8.10.0) (2020-06-17)
-
-### Bug Fixes
-
-- **plugins/plugin-sandbox:** incorrect package.json name ([a9d2243](https://github.com/IBM/kui/commit/a9d2243)), closes [#3854](https://github.com/IBM/kui/issues/3854)
-- core defaultContext logic assumes existence if @kui-shell/client ([a8a32cf](https://github.com/IBM/kui/commit/a8a32cf)), closes [#3849](https://github.com/IBM/kui/issues/3849)
-
-### Features
-
-- react helpers ([f6bea1f](https://github.com/IBM/kui/commit/f6bea1f))
-
-### BREAKING CHANGES
-
-- this PR removes plugins/plugin-client-default
-
-# [8.9.0](https://github.com/IBM/kui/compare/v4.5.0...v8.9.0) (2020-06-09)
-
-### Bug Fixes
-
-- **plugins/plugin-sandbox:** incorrect package.json name ([a9d2243](https://github.com/IBM/kui/commit/a9d2243)), closes [#3854](https://github.com/IBM/kui/issues/3854)
-- core defaultContext logic assumes existence if @kui-shell/client ([a8a32cf](https://github.com/IBM/kui/commit/a8a32cf)), closes [#3849](https://github.com/IBM/kui/issues/3849)
-
-### Features
-
-- react helpers ([f6bea1f](https://github.com/IBM/kui/commit/f6bea1f))
-
-### BREAKING CHANGES
-
-- this PR removes plugins/plugin-client-default
-
-# [8.7.0](https://github.com/IBM/kui/compare/v4.5.0...v8.7.0) (2020-05-08)
-
-### Bug Fixes
-
-- **plugins/plugin-sandbox:** incorrect package.json name ([a9d2243](https://github.com/IBM/kui/commit/a9d2243)), closes [#3854](https://github.com/IBM/kui/issues/3854)
-- core defaultContext logic assumes existence if @kui-shell/client ([a8a32cf](https://github.com/IBM/kui/commit/a8a32cf)), closes [#3849](https://github.com/IBM/kui/issues/3849)
-
-### Features
-
-- react helpers ([f6bea1f](https://github.com/IBM/kui/commit/f6bea1f))
-
-### BREAKING CHANGES
-
-- this PR removes plugins/plugin-client-default
-
-## [8.6.1](https://github.com/IBM/kui/compare/v4.5.0...v8.6.1) (2020-04-25)
-
-### Bug Fixes
-
-- **plugins/plugin-sandbox:** incorrect package.json name ([a9d2243](https://github.com/IBM/kui/commit/a9d2243)), closes [#3854](https://github.com/IBM/kui/issues/3854)
-- core defaultContext logic assumes existence if @kui-shell/client ([a8a32cf](https://github.com/IBM/kui/commit/a8a32cf)), closes [#3849](https://github.com/IBM/kui/issues/3849)
-
-### Features
-
-- react helpers ([f6bea1f](https://github.com/IBM/kui/commit/f6bea1f))
-
-### BREAKING CHANGES
-
-- this PR removes plugins/plugin-client-default
-
-# [8.6.0](https://github.com/IBM/kui/compare/v4.5.0...v8.6.0) (2020-04-23)
-
-### Bug Fixes
-
-- **plugins/plugin-sandbox:** incorrect package.json name ([a9d2243](https://github.com/IBM/kui/commit/a9d2243)), closes [#3854](https://github.com/IBM/kui/issues/3854)
-- core defaultContext logic assumes existence if @kui-shell/client ([a8a32cf](https://github.com/IBM/kui/commit/a8a32cf)), closes [#3849](https://github.com/IBM/kui/issues/3849)
-
-### Features
-
-- react helpers ([f6bea1f](https://github.com/IBM/kui/commit/f6bea1f))
-
-### BREAKING CHANGES
-
-- this PR removes plugins/plugin-client-default
-
-# [8.5.0](https://github.com/IBM/kui/compare/v4.5.0...v8.5.0) (2020-04-19)
-
-### Bug Fixes
-
-- **plugins/plugin-sandbox:** incorrect package.json name ([a9d2243](https://github.com/IBM/kui/commit/a9d2243)), closes [#3854](https://github.com/IBM/kui/issues/3854)
-- core defaultContext logic assumes existence if @kui-shell/client ([a8a32cf](https://github.com/IBM/kui/commit/a8a32cf)), closes [#3849](https://github.com/IBM/kui/issues/3849)
-
-### Features
-
-- react helpers ([f6bea1f](https://github.com/IBM/kui/commit/f6bea1f))
-
-### BREAKING CHANGES
-
-- this PR removes plugins/plugin-client-default
-
-## [8.4.2](https://github.com/IBM/kui/compare/v4.5.0...v8.4.2) (2020-04-10)
-
-### Bug Fixes
-
-- **plugins/plugin-sandbox:** incorrect package.json name ([a9d2243](https://github.com/IBM/kui/commit/a9d2243)), closes [#3854](https://github.com/IBM/kui/issues/3854)
-- core defaultContext logic assumes existence if @kui-shell/client ([a8a32cf](https://github.com/IBM/kui/commit/a8a32cf)), closes [#3849](https://github.com/IBM/kui/issues/3849)
-
-### Features
-
-- react helpers ([f6bea1f](https://github.com/IBM/kui/commit/f6bea1f))
-
-### BREAKING CHANGES
-
-- this PR removes plugins/plugin-client-default
-
-## [8.4.1](https://github.com/IBM/kui/compare/v4.5.0...v8.4.1) (2020-04-10)
-
-### Bug Fixes
-
-- **plugins/plugin-sandbox:** incorrect package.json name ([a9d2243](https://github.com/IBM/kui/commit/a9d2243)), closes [#3854](https://github.com/IBM/kui/issues/3854)
-- core defaultContext logic assumes existence if @kui-shell/client ([a8a32cf](https://github.com/IBM/kui/commit/a8a32cf)), closes [#3849](https://github.com/IBM/kui/issues/3849)
-
-### Features
-
-- react helpers ([f6bea1f](https://github.com/IBM/kui/commit/f6bea1f))
-
-### BREAKING CHANGES
-
-- this PR removes plugins/plugin-client-default
-
-# [8.4.0](https://github.com/IBM/kui/compare/v4.5.0...v8.4.0) (2020-04-10)
-
-### Bug Fixes
-
-- **plugins/plugin-sandbox:** incorrect package.json name ([a9d2243](https://github.com/IBM/kui/commit/a9d2243)), closes [#3854](https://github.com/IBM/kui/issues/3854)
-- core defaultContext logic assumes existence if @kui-shell/client ([a8a32cf](https://github.com/IBM/kui/commit/a8a32cf)), closes [#3849](https://github.com/IBM/kui/issues/3849)
-
-### Features
-
-- react helpers ([f6bea1f](https://github.com/IBM/kui/commit/f6bea1f))
-
-### BREAKING CHANGES
-
-- this PR removes plugins/plugin-client-default
-
-# [8.1.0](https://github.com/IBM/kui/compare/v4.5.0...v8.1.0) (2020-04-04)
-
-### Bug Fixes
-
-- **plugins/plugin-sandbox:** incorrect package.json name ([a9d2243](https://github.com/IBM/kui/commit/a9d2243)), closes [#3854](https://github.com/IBM/kui/issues/3854)
-- core defaultContext logic assumes existence if @kui-shell/client ([a8a32cf](https://github.com/IBM/kui/commit/a8a32cf)), closes [#3849](https://github.com/IBM/kui/issues/3849)
-
-### Features
-
-- react helpers ([f6bea1f](https://github.com/IBM/kui/commit/f6bea1f))
-
-### BREAKING CHANGES
-
-- this PR removes plugins/plugin-client-default
-
-# [8.0.0](https://github.com/IBM/kui/compare/v4.5.0...v8.0.0) (2020-03-20)
-
-### Bug Fixes
-
-- **plugins/plugin-sandbox:** incorrect package.json name ([a9d2243](https://github.com/IBM/kui/commit/a9d2243)), closes [#3854](https://github.com/IBM/kui/issues/3854)
-- core defaultContext logic assumes existence if @kui-shell/client ([a8a32cf](https://github.com/IBM/kui/commit/a8a32cf)), closes [#3849](https://github.com/IBM/kui/issues/3849)
-
-### Features
-
-- react helpers ([f6bea1f](https://github.com/IBM/kui/commit/f6bea1f))
-
-### BREAKING CHANGES
-
-- this PR removes plugins/plugin-client-default
