--- plugins/plugin-sandbox/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "@kui-shell/plugin-sandbox",
-  "version": "8.10.3",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "Apache-2.0",
-  "private": true,
-  "publishConfig": {
-    "access": "public"
-  }
-}
