--- plugins/plugin-sandbox/prescan.json
@@ -1,150 +1,0 @@
-{
-  "preloads": [
-    {
-      "route": "plugin-core-support",
-      "path": "plugin-core-support/dist/preload.js"
-    }
-  ],
-  "themeSets": [
-    {
-      "themes": [
-        {
-          "name": "Light",
-          "css": "light.css",
-          "description": "Default Light",
-          "style": "light",
-          "apiVersion": "v2"
-        },
-        {
-          "name": "Dark",
-          "css": "dark.css",
-          "description": "Default Dark",
-          "style": "dark",
-          "apiVersion": "v2"
-        },
-        {
-          "name": "Atelier",
-          "css": "atelier-seaside.css",
-          "style": "dark",
-          "apiVersion": "v2"
-        },
-        {
-          "name": "Dracula",
-          "css": "dracula.css",
-          "style": "dark",
-          "apiVersion": "v2"
-        },
-        {
-          "name": "Gruvbox",
-          "css": "gruvbox-dark-medium.css",
-          "style": "dark",
-          "apiVersion": "v2"
-        },
-        {
-          "name": "Lucario",
-          "css": "lucario.css",
-          "style": "dark",
-          "apiVersion": "v2"
-        },
-        {
-          "name": "Monokai",
-          "css": "monokai.css",
-          "style": "dark",
-          "apiVersion": "v2"
-        },
-        {
-          "name": "Nord",
-          "css": "nord.css",
-          "style": "dark",
-          "apiVersion": "v2"
-        },
-        {
-          "name": "Paraiso",
-          "css": "paraiso.css",
-          "style": "dark",
-          "apiVersion": "v2"
-        },
-        {
-          "name": "Solarized",
-          "css": "solarized-dark.css",
-          "style": "dark",
-          "apiVersion": "v2"
-        },
-        {
-          "name": "tomorrow-night",
-          "css": "tomorrow-night.css",
-          "description": "Tomorrow Night",
-          "style": "dark",
-          "apiVersion": "v2"
-        }
-      ],
-      "plugin": "plugin-core-themes"
-    }
-  ],
-  "commandToPlugin": {
-    "/confirm": "plugin-client-common",
-    "/shell": "plugin-core-support",
-    "/quit": "plugin-core-support",
-    "/exit": "plugin-core-support",
-    "/clear": "plugin-core-support",
-    "/base64": "plugin-core-support",
-    "/prompt": "plugin-core-support",
-    "/sleep": "plugin-core-support",
-    "/history": "plugin-core-support",
-    "/!!": "plugin-core-support",
-    "/again": "plugin-core-support",
-    "/tab/switch": "plugin-core-support",
-    "/tab/new": "plugin-core-support",
-    "/tab/close": "plugin-core-support",
-    "/zoom/get": "plugin-core-support",
-    "/zoom/set": "plugin-core-support",
-    "/zoom/reset": "plugin-core-support",
-    "/window": "plugin-core-support",
-    "/window/help": "plugin-core-support",
-    "/window/max": "plugin-core-support",
-    "/window/unmax": "plugin-core-support",
-    "/window/close": "plugin-core-support",
-    "/theme/list": "plugin-core-support",
-    "/themes": "plugin-core-support",
-    "/theme": "plugin-core-support",
-    "/theme/set": "plugin-core-support",
-    "/theme/current": "plugin-core-support",
-    "/theme/reset": "plugin-core-support",
-    "/version": "plugin-core-support/about",
-    "/about": "plugin-core-support/about",
-    "/getting/started": "plugin-core-support/about",
-    "/edit": "plugin-editor",
-    "/plugin/commands": "plugin-manager",
-    "/plugin/compile": "plugin-manager",
-    "/plugin/install": "plugin-manager",
-    "/plugin/get": "plugin-manager",
-    "/plugin/list": "plugin-manager",
-    "/plugin/home": "plugin-manager",
-    "/plugin/remove": "plugin-manager",
-    "/plugin/uninstall": "plugin-manager",
-    "/plugin/version": "plugin-manager"
-  },
-  "topological": {},
-  "flat": [
-    {
-      "route": "plugin-client-common",
-      "path": "plugin-client-common/dist/plugin.js"
-    },
-    {
-      "route": "plugin-core-support",
-      "path": "plugin-core-support/dist/plugin.js"
-    },
-    {
-      "route": "plugin-core-support/about",
-      "path": "plugin-core-support/about/dist/plugin.js"
-    },
-    {
-      "route": "plugin-editor",
-      "path": "plugin-editor/dist/plugin.js"
-    }
-  ],
-  "overrides": {},
-  "usage": {},
-  "catchalls": [],
-  "docs": {}
-}
